- feat: theme 내부로 media query 이동 및 theme type 적용.
