+++ app/styles/styledComponents.d.ts
@@ -1,0 +1,6 @@
+import 'styled-components';
+import { ThemeType } from './theme';
+
+declare module 'styled-components' {
+  export interface DefaultTheme extends ThemeType {}
+}
