+++ app/styles/theme.ts
@@ -1,5 +1,25 @@
 import { css } from 'styled-components';
 
+export interface DeviceProps {
+  mobile: number;
+  tablet: number;
+  desktop: number;
+}
+
+const deviceSizes: DeviceProps = {
+  mobile: 767,
+  tablet: 991,
+  desktop: 992,
+};
+
+const mediaQuery = {
+  mobile: `@media only screen and (max-width: ${deviceSizes.mobile}px)`,
+  tablet: `@media only screen and (max-width: ${deviceSizes.tablet}px) and (min-width: ${
+    deviceSizes.mobile + 1
+  }px)`,
+  desktop: `@media only screen and (max-width: ${deviceSizes.desktop}px)`,
+};
+
 const color = {
   static: {
     dark: '#000000',
@@ -155,6 +175,8 @@ const font = {
   `,
 };
 
-const theme = { color, font };
+const theme = { color, font, mediaQuery };
 
 export default theme;
+
+export type ThemeType = typeof theme;
